Update README.md
The purpose of Stevens Tip Calculator is too figure out the correct customary tip to be left in his region.The calculator os useful for quick use, accuracy, and is user friendly. To calculate an exact tip you would need a calculator and to do the math yourself. This does not require that. In addition, this project gave me a lot of experience and practice with coding.